src: get rid of controller
