+++ src/kube-red.ts
@@ -1,5 +1,5 @@
 import { NodeDef, NodeAPI, Node, NodeMessageInFlow } from "node-red";
-import registerLowerCase from "./lower-case/controller";
+import registerLowerCase from "./lower-case/node";
 
 export default function (RED: NodeAPI) {
     registerLowerCase(RED);
